Generalize creation of dynamic_reconfigure servers based on Settings
In upcoming work we will add support for Settings2D. The Settings2D API
in ROS will be analogous to the 3D API.

In this commit some preparation work is done to facilitate this:

- Generalize the old DRFrameConfig struct. This was previously used for
only 3D frame settings to setup the dynamic_reconfigure server, and store
the Config object each time the config is changed by the client.
DRFrameConfig is changed to templated struct ConfigDRServer, and is also
used for 3D common settings. In upcoming work this will also be used for
2D frame settings.

- Generalize/template the functions that sets up these servers (setting
the min, max and default value of all the settings, as well as the
callback). This is now template function setupConfigDRServer, which
ensures that all the dynamic_reconfigure servers are setup the same way.

- To faciliate the point above, the build-time generated methods that
convert from a Zivid::Settings object to ROS CaptureGeneralConfig /
CaptureFrameConfig objects have been made template methods:
getMinConfigFromZividSettings, getMaxConfigFromZividSettings,
getDefaultConfigFromZividSettings. The ROS ConfigType
(CaptureGeneralConfig/CaptureFrameConfig) is the template parameter.

- The settings generator also no longer hardcodes "Zivid::Settings"
as the class name, but gets this name via Zivid::Settings::name.