Add support for 2D capture
Add a new service, capture_2d which triggers 2D capture and publishes the
result on the color/image_color topic.

Configuration of 2D settings is done similar to 3D captures, using
dynamic_reconfigure. The 2D settings are at /capture_2d/frame_0. 2D HDR
(multiple frames) is not something the SDK supports currently, but the API
is designed to support it in the future by adding more
/capture_2d/frame_<n> nodes.

C++ and Python samples are also added.