Improvements to ConfigDRServer initialization
Previously, ConfigDRServer was a struct which just held some public
members, and there was a templated function setupConfigDRServer to
initialize an instance of this struct. Change to full-blown class and
initialize it fully in the constructor. This makes this class more
contained.

A motivation for this change is that we want to limit access to the class
members. By directly accessing the members it is possible to get into bad
state, for example by updating the locally cached config but not updating
the dynamic_reconfigure server, or opposite, update server but not local
cached config.