test_zivid_camera: Fix broken test (add missing short sleep)
This test started to fail consistently in the CI around May 7th 2020.
Unclear why it worked prior to this date, but probably a timing issue.
Adding a short sleep in testCapturePoints (like we do all other places
in the tests) fixes the issue. The sleep is necessary to let ROS
messages propagate.

We still have an issue with the Intel OpenCL CPU driver (see
issue #24) which will be fixed in an upcoming commit.