Fix broken Azure CI status badge
In commit 8b53a4d3f7e6 the CI status badge image was changed to using
the wrong project and defintion id's. We incorrectly used zivid-python
id's.

In this commit the correct project id and definition id for zivid-ros
are used. The image URL was created following the guide in
https://shields.io/category/build.

The badge link previously gave 404. It is now updated to point to the
master branch.