test_zivid_camera: Use AsyncSpinner in unit tests
Previously the ROS event loop was spinned in the main thread, using
helper method sleepAndSpin. However, when using
dynamic_reconfigure::Client::getCurrentConfiguration it will block the
thread until a configuration has been received (or it gives up and
returns false after a timeout has been reached). When the thread was
blocked no further events was processed. This means we were unable to
get current configuration and subsequent  tests would fail. To work
around this strategic sleepAndSpin were needed to ensure the event loop
was processed before getCurrentConfiguration was called.

In this commit the ROS event loop is moved to a different thread, using
ros::AsyncSpinner. This means that the event loop will continue
processing even if the main (test) thread is blocking. This makes it
easier to write tests that are stable. We no longer need to spin the
event loop manually within the tests.