Fix CI failure on Ubuntu 16 due to ROS kinetic being EOL
The CI pipeline for U16 was failing with the error:
zivid_camera: Cannot locate rosdep definition for [image_transport]

The problem is that kinetic has been EOL since May 14th, so rosdep update
would not include this distro. By explicitly setting the --rosdistro argument
in rosdep update we will ensure the current distro is included. It
should also speed up the CI a tiny bit, since we are not updating other
distros.

Fix for issue 43.