Add topic for normals
Normals are now published via topic normals/xyz. The data is published
as a PointCloud2 that contains only normal_{x, y, z} fields, without any
additional padding fields. The normals are only calculated and copied
if anyone is actually subscribed to the normals/xyz topic.

In the future we may get requests for a topic that contains normals
together with points, similar to how we today provide points/xyzrgb.
In that case, we should provide that output format natively in the SDK
first so that it can be used in ROS without for loops.

Normals were introduced in Zivid SDK 2.5, so this commit bumps the
minimum required SDK from 2.0 to 2.5.