Add service to capture and save frame
It is very valuable to be able to save the frame
as a ZDF file, or export to other formats.
This commit introduces the service capture_and_save_frame
which both captures and saves the frame. It takes a
path as an argument. See README.md#capture_and_save_frame
for more information.

In order to show how to use this service we also add a
sample, both for Python and C++.

Closes #65