test_zivid_camera: Add separate test that checks zivid_camera node ready
Before the zivid_camera tests can run, the zivid_camera driver node that
the  test connects to must be up and running. The startup of this driver
happens in parallel with the test being invoked, and the startup can take
some time, so that is why we have the waitForReady() helper function.

In this commit, I do two changes:

1) Add an initial test that waits for the node to be ready. Previously,
any error during the startup would be instead reported as failure in
testServiceCameraInfoModelName. It is better to have an explicit
test that the camera node is ready.

2) I remove waitForReady() calls for all subsequent tests. This is not
necessary as the node is already ready at this point. All of the test
cases are testing towards the same zivid_camera instance.