test_zivid_camera.cpp: Do not test for exact count of NaN's
This checked that the number of NaN's in the point cloud is exactly X.
However, X can vary a tiny bit from compute device/driver etc. Below
in this test we compare the ROS output from the output from capturing
with Zivid SDK directly. This should be good enough for testing the
ROS driver specifically, we don't need to test the SDK here.

We keep a test that the point cloud is not fully NaN or fully non-NaN.