Make python
This was an attempt at making the ROS2 launch files work with the Python
samples on Windows. The issue is that, if we simply try to launch the
python script directly, we get the following error:

```
OSError: [WinError 193] %1 is not a valid Win32 application
```

In this commit, we try to make a Python module out of the samples,
install them using `ament_cmake_python`, and use a `setup.cfg` to add a
console script entry point. This way, we can launch the samples using
both `ros2 run` and `ros2 launch` commands.

This approach was modeled after: https://github.com/ros/xacro/pull/304

This seemed to work. However, a new issue emerged. While `ros2 run`
still works fine, now we are not getting any log output when launching
it with `ros2 launch`. We seem to be encountering this issue:

MISC-2024-06-19-ros2-testing