Fix building with other compilers
The flag `CMAKE_CXX_COMPILER` was not previously passed to CMake due
to the double quotes in the `build.sh` script.

This implies that we didn't previously build with clang. Building
with clang started emitting some warnings, causing errors. These
warnings have been addressed here as well.

Also removed warning `return-std-move-in-c++11` as this was removed in
Clang 13 and now causing a compile error:
https://releases.llvm.org/13.0.0/tools/clang/docs/ReleaseNotes.html