Fix use of uninitialized memory in zivid_camera
The constructor of CaptureSettingsController fires the "set parameters"
event on the camera node. This is then propagated by the registered
callback to the settings controllers. This causes UB because one of the
controllers is uninitialized. This UB results in a crash on Humble.

To fix this we initialize the settings controllers to nullptr before the
callback is registered.