Do not terminate the node if capturing with no settings set
Specifying settings both via yml and file, or not specifying settings
at all, is considered an error in the ROS driver. Prior to this commit,
when the user invoked capture in any of these scenarios, then we threw
an exception that we let propagate, so that the ROS node would terminate
with an error like:

```
terminate called after throwing an instance of 'std::runtime_error'
  what():  Both 'settings_file_path' and 'settings_yaml' parameters are empty! Please set one of these parameters.
[ros2run]: Aborted
```

Other errors/exceptions that could occurr, such as if Zivid SDK's
capture() function threw an exception, would be returned as a
TriggerResponse with "success" false, and "message" containing the
exception string. In this case the camera node continues to live,
so the user can fix the settings, and then call capture again.

The existing behavior was intentional; the idea was that what we
considered a programming/setup mistake, specifically capturing without
having a settings yaml string or file path configured, the exception
should be propagated fully, which is what causes the ROS2 node to
die. However after some consideration, I think this is overly strict,
gives little benefit, and is annoying for the user when they are
trying to configure the driver correctly.

See https://github.com/zivid/zivid-ros/issues/116

Fixes part of issue 116