Do not terminate the node if capture_assistant/suggest_settings throws
Prior to this commit, an exception during the
"capture_assistant/suggest_settings" command would cause the
zivid_camera driver to throw an exception, which then would cause the
node to terminate.

In this commit, we instead catch the exceptions that would occurr, and
pass error status + message to the caller. See parent commit for the
reasoning around this change.

This requires us to add "success" and "message" fields to the
CaptureAssistantSuggestSettings.srv response message. This is modelled
after the standard Trigger message,
https://docs.ros.org/en/noetic/api/std_srvs/html/srv/Trigger.html,
which is the one we are using for capture commands. We can't use
std_srvs/Trigger directly for Capture Assistant, since CA has input
parameters max_capture_time and ambient_light_frequency.

Fixes issue #116