Add URDF/Xacro for Zivid cameras
Add a new package `zivid_description` containing URDF/Xacro files for
Zivid cameras.

Add new launch files to the sample package to launch a Zivid camera with
the URDF. Modify existing launch files to also load the URDF.
Additionally, change the Zivid RViz configuration to use the robot
description.

This way, when launching the Zivid camera with RViz, the camera will
appear in the scene. Point clouds published from captures will be shown
in the RViz scene appropriately placed in relation to the camera.

This commit builds on top of pull request #17. The following changes are
made here compared to pull request #17:

- Convert to ROS2.
- Add new visual and collision mesh for Zivid 2/2+/2+R cameras.
- Add FOV frustum meshes for the different camera models (toggleable).
- Add physical parameters for the Zivid camera models, for appropriate
  placement of the point cloud in relation to the camera.
- Remove Zivid One+ support.
- Make Xacro error out with a reasonable error message when specifying
  an unknown model.
- Rename argument type -> model.
- Move launch files to `zivid_samples`.