Implement infield correction
Add support for infield correction in the zivid_camera ROS driver.
Please see the readme for usage instructions.

This commit also adds a new sample to demonstrate the new functionality.

Additionally, an RViz panel is added for interactive infield correction,
with the new `zivid_rviz_plugin` package.

New tests are added. A new file camera from the Zivid sample data is
used to test the new feature with a calibration board.