Add `acquisition_done` topic to zivid_camera
When using the Zivid SDK directly there is information in when the
capture API returns. This indicates that the camera is done with
acquisitions. At this point it is safe to move the scene (e.g. robot).
In ROS the capture service wraps the capture call, but also extraction
and processing of the point cloud. The same goes for 2D capture.

Use this topic to control when to, e.g., move the robot.